[FIX] mail: multiple action_unfollow on livechat leave
Before this commit, leaving a livechat channel from the Discuss app would
trigger multiple `action_unfollow` actions:
- One initiated by the Discuss app.
- Another triggered by the `channel/leave` notification.

This commit resolves the issue by ensuring that the `channel/leave` notification
does not redundantly trigger an `action_unfollow`.

Additionally, this commit addresses minor code structure issues.

co-authored by: Alexandre Kühn (aku) <aku@odoo.com>
