+++ addons/im_livechat/static/src/core/common/chat_window_model_patch.js
@@ -13,16 +13,20 @@ const chatWindowPatch = {
         super.setup(...arguments);
         this.livechatStep = CW_LIVECHAT_STEP.NONE;
     },
-    close(...args) {
+    close(options = {}) {
         if (this.thread?.channel_type !== "livechat") {
-            return super.close(...args);
+            return super.close(...arguments);
+        }
+        if (options.force) {
+            this.livechatStep = CW_LIVECHAT_STEP.NONE;
+            return super.close(...arguments);
         }
         const isSelfVisitor = this.thread.livechatVisitorMember?.persona?.eq(this.store.self);
         switch (this.livechatStep) {
             case CW_LIVECHAT_STEP.NONE: {
                 if (this.thread.isTransient) {
                     this.thread.delete();
-                    super.close(...args);
+                    super.close(...arguments);
                     break;
                 }
                 if (!this.thread.livechat_active) {
@@ -30,7 +34,7 @@ const chatWindowPatch = {
                         this.livechatStep = CW_LIVECHAT_STEP.FEEDBACK;
                         this.open({ focus: true, notifyState: this.thread?.state !== "open" });
                     } else {
-                        super.close(...args);
+                        super.close(...arguments);
                     }
                     break;
                 }
@@ -48,31 +52,18 @@ const chatWindowPatch = {
                     this.livechatStep = CW_LIVECHAT_STEP.FEEDBACK;
                 } else {
                     this.livechatStep = CW_LIVECHAT_STEP.NONE;
-                    super.close(...args);
+                    super.close(...arguments);
                 }
                 break;
             }
             case CW_LIVECHAT_STEP.FEEDBACK: {
-                super.close(...args);
+                super.close(...arguments);
                 break;
             }
         }
         if (this.livechatStep !== CW_LIVECHAT_STEP.CONFIRM_CLOSE) {
             this.store.env.services["im_livechat.livechat"]?.leave();
             this.store.env.services["im_livechat.chatbot"]?.stop();
-        }
-    },
-    _onClose() {
-        if (
-            this.thread?.channel_type === "livechat" &&
-            this.thread.livechatVisitorMember?.persona?.notEq(this.store.self)
-        ) {
-            const thread = this.thread; // save ref before delete
-            super._onClose();
-            this.delete();
-            thread.leaveChannel({ force: true });
-        } else {
-            super._onClose();
         }
     },
 };
